+++ tests/test_tlf.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # flake8: noqa
 
 import pytest
@@ -22,13 +21,13 @@ from . import requires_xarray  # noqa
 from . import requires_xesmf  # noqa
 
 
-def test_compute_weighted_means_ds():
+def test_compute_weighted_means_ds_gdf():
     netcdffile = xw.test_netcdf[0]
-    shp = xw.get_region("states")
+    gdf = xw.get_region("states")
     ds = xr.open_dataset(netcdffile)
     xw.compute_weighted_means_ds(
         ds,
-        shp,
+        gdf=gdf,
         time_range=["2007-01-01", "2007-11-30"],
         column_names=[
             "institute_id",
@@ -41,14 +40,58 @@ def test_compute_weighted_means_ds():
             "standard_name",
             "not_available",
         ],
+        output=".",
     )
 
 
-def test_compute_weighted_means():
+def test_compute_weighted_means_ds_region():
     netcdffile = xw.test_netcdf[0]
-    xw.compute_weighted_means(
-        netcdffile,
+    ds = xr.open_dataset(netcdffile)
+    xw.compute_weighted_means_ds(
+        ds,
         region="states",
+        time_range=["2007-01-01", "2007-11-30"],
+        column_names=[
+            "institute_id",
+            "driving_model_id",
+            "experiment_id",
+            "driving_model_ensemlbe_member",
+            "model_id",
+            "rcm_version_id",
+            "units",
+            "standard_name",
+            "not_available",
+        ],
+        output="test",
+    )
+
+
+def test_compute_weighted_means_region():
+    netcdffile = xw.test_netcdf[0]
+    ds = xr.open_dataset(netcdffile)
+    dataset_dict = {"test_data": ds}
+    xw.compute_weighted_means(
+        dataset_dict,
+        region="states",
+        subregion=[
+            "01_Schleswig-Holstein",
+            "02_Hamburg",
+            "03_Niedersachsen",
+            "04_Bremen",
+        ],
+        merge_columns=["all", "NorthSeaCoast"],
+        output="test.csv",
+    )
+
+
+def test_compute_weighted_means_gdf():
+    netcdffile = xw.test_netcdf[0]
+    ds = xr.open_dataset(netcdffile)
+    dataset_dict = {"test_data": ds}
+    gdf = xw.get_region("states")
+    xw.compute_weighted_means(
+        dataset_dict,
+        gdf=gdf,
         subregion=[
             "01_Schleswig-Holstein",
             "02_Hamburg",