+++ xweights/_weightings.py
@@ -1,16 +1,14 @@
 import xarray as xr
 import xesmf as xe
 
-from ._domains import get_domain
-
 
 def get_spatial_averager(ds, gdf, name=None):
     """get xesmf's spatial averager
 
     Parameters
     ----------
-    ds: xr.Dataset or str
-        Dataset or name of CORDEX_domain
+    ds: xr.Dataset
+
     gdf:
         gp.GeoDataFrame
     name: str (optional)
@@ -20,8 +18,6 @@ def get_spatial_averager(ds, gdf, name=None):
     -------
     savg - xesmf.SpatialAverager
     """
-    if isinstance(ds, str):
-        ds = get_domain(ds)
     savg = xe.SpatialAverager(ds, gdf.geometry)
     if name is None:
         name = gdf.attrs["name"]
@@ -30,14 +26,14 @@ def get_spatial_averager(ds, gdf, name=None):
     return savg
 
 
-def spatial_averaging(ds, shp=None, savg=None):
+def spatial_averaging(ds, gdf=None, savg=None):
     """xesmf's spatial averager
 
     Parameters
     ----------
     ds: xr.Dataset
 
-    shp: gp.GeoDataFrame (optional)
+    gdf: gp.GeoDataFrame (optional)
 
     savg: xesmf.SpatialAverager (optional)
 
@@ -45,7 +41,7 @@ def spatial_averaging(ds, shp=None, savg=None):
     -------
     out - xr.Dataset
         Dataset containing a time series of spatial averages
-        for each geometry in ``shp``
+        for each geometry in ``gdf``
 
     Example
     -------
@@ -54,22 +50,22 @@ def spatial_averaging(ds, shp=None, savg=None):
         import xweights as xw
         import xarray as xr
 
-        netcdffile = ("/work/kd0956/CORDEX/data/cordex/output/EUR-11/CLMcom/"
-                     "MIROC-MIROC5/rcp85/r1i1p1/CLMcom-CCLM4-8-17/v1/mon/tas/"
-                     "v20171121/tas_EUR-11_MIROC-MIROC5_rcp85_r1i1p1_"
-                     "CLMcom-CCLM4-8-17_v1_mon_200601-201012.nc")
+        ncf = (
+            "tas_EUR-11_MIROC-MIROC5_rcp85_r1i1p1_"
+            "CLMcom-CCLM4-8-17_v1_mon_200601-201012.nc"
+        )
 
-        ds = xr.open_dataset(netcdffile)
+        ds = xr.open_dataset(ncf)
 
         shp = xw.get_region('states')
 
-        out = xw.spatial_averager(ds, shp)
+        out = xw.spatial_averager(ds, gdf)
 
     """
     if savg is None:
-        savg = get_spatial_averager(ds, shp)
+        savg = get_spatial_averager(ds, gdf)
     elif isinstance(savg, str):
-        savg = get_spatial_averager(savg, shp)
+        savg = get_spatial_averager(savg, gdf)
 
     nnz = [w.data.nnz for w in savg.weights]
     out = savg(ds)